eweather: enable usage of c++, add 2 formats string to display the weather temp in the plugin. Maybee need to check it from theme too ?
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@58275 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ PROTO/libeweather/src/lib/EWeather_Plugins.h
@@ -16,9 +16,9 @@ typedef void (*Plugin_Code_Updated) (EWeather *eweather);
 struct EWeather_Data
 {
    EWeather_Type type;
-   int temp;
-   int temp_min;
-   int temp_max;
+   double temp;
+   double temp_min;
+   double temp_max;
    char city[256];
    char region[256];
    char country[256];
